move logging, debug_settings and ota_service from app.configs to separate configs package
